+++ mydatastore.cpp
@@ -115,7 +115,7 @@ void myDataStore::addToCart(User* u, Product* p) {
 void myDataStore::viewCart(User* u) {
 	int itemNo = 1;
 	vector<Product*> temp = carts_[u];
-	for (int i = 0; i < temp.size(); i++) {
+	for (int i = 0; i < (int) temp.size(); i++) {
 		cout << "Item " << itemNo << endl;
 		temp[i]->displayString();
 		itemNo++;
@@ -145,7 +145,7 @@ void myDataStore::buyCart(User* u) {
 
 	// 	}
 	// }
-	
+
 	vector<Product*>::iterator it = carts_[u].begin();
 	while (it != carts_[u].end()) {
 		productPrice = (*it)->getPrice();
